# pages/swagger-config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xavier Api",
    "description": "Welcome to the Xavier Api, your gateway to streamlined and efficient application integration. Designed with simplicity and versatility in mind, Xavier Api empowers developers to effortlessly manage and interact with essential data and functionalities. Whether you're building a web application, mobile app, or integrating with other services, Laurine Api provides a robust foundation for seamless operations.",
    "contact": {
      "email": "https://t.me/FVendettaV"
    },
    "license": {
      "name": "API Development Information",
      "url": "https://whatsapp.com/channel/0029VbACugX6rsQyHw0Wvr1g"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://japi.vercel.app",
      "description": "Xavier Api"
    }
  ],
  "paths": {
...
    "/api/ai/mixtral-8x7b-32768": {
      "get": {
        "tags": ["Artificial Intelligence"],
        "description": "mixtral-8x7b-32768 is a Mixture of Experts (MoE) model by Mistral AI, featuring 8 experts, each with 7 billion parameters, with only 2 experts active per token. The number 32768 represents the supported context length, allowing for longer and more efficient text processing compared to large monolithic models.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/ai/deepseek": {
      "get": {
        "tags": ["Artificial Intelligence"],
        "description": "DeepSeek is an artificial intelligence (AI) model known for its efficiency and cost-effectiveness. It is designed to generate text-based responses and solutions, and has been used in various applications, including chatbots and text-based problem-solving. DeepSeek offers a competitive alternative to other AI models due to its performance and affordability.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/ai/gemma2-9b-it": {
      "get": {
        "tags": ["Artificial Intelligence"],
        "description": "Gemma 2-9B-IT is an AI model developed by Google with 9 billion parameters, fine-tuned for instruction following. It is based on a decoder-only architecture, trained on 8 trillion tokens, and is suitable for tasks such as text generation, question answering, and natural language processing.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/ai/heckai": {
      "get": {
        "tags": ["Artificial Intelligence"],
        "description": "HeckAI is an artificial intelligence (AI) system developed specifically to predict the outcomes of chemical reactions, particularly in Mizoroki-Heck C-glycosylation. The system applies machine learning algorithms designed to operate with limited data by leveraging knowledge-based chemical descriptors. With this approach, HeckAI can quickly and efficiently provide reaction yield predictions, thereby helping researchers optimize reaction designs without the need for extensive laboratory experiments.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/downloader/douyindl": {
      "get": {
        "tags": ["Downloader"],
        "description": "Douyin Downloader is a tool or application that allows users to download videos from Douyin, the Chinese version of TikTok. This tool enables users to save videos in high quality, often without a watermark.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/downloader/fbdownload": {
      "get": {
        "tags": ["Downloader"],
        "description": "Facebook Downloader is a tool or app used to download videos or other content from Facebook. It allows users to save videos, images, or audio to their device for offline viewing. It’s simple to use, enabling quick downloads with just a few steps. However, users should be mindful of copyright rules and Facebook's guidelines when using such tools.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/downloader/spotifydl": {
      "get": {
        "tags": ["Downloader"],
        "description": "Spotify Downloader is a website or tool that allows users to easily and quickly download songs from Spotify. With this service, users can save their favorite songs to listen to offline anytime without needing an internet connection. Additionally, SPOTIFYDL offers easy access without the need for extra apps, making it more practical and flexible.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/cai/bocchi": {
      "get": {
        "tags": ["Character AI"],
        "description": "Bocchi Hitori, also known as Bocchi, is the main character of the anime Bocchi the Rock! She is a shy and socially awkward girl who dreams of becoming a famous guitarist. Despite her amazing guitar skills, she struggles with social interactions and often experiences anxiety. However, after joining Kessoku Band, she slowly starts stepping out of her comfort zone and learning to work with her friends.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}